added alot of stuff. add page kind of working. need a way to add genres to books.

+++ Actual Project/CarouselForBooksSolution/CarouselForBooksApplication/Controllers/BookController.cs
@@ -11,8 +11,8 @@ namespace CarouselForBooksApplication.Controllers
 {
     public class BookController : Controller
     {
-        private readonly IRepo<int, Book> _repo;
-        public BookController(IRepo<int, Book> repo)
+        private readonly IBook<int, Book, string> _repo;
+        public BookController(IBook<int, Book, string> repo)
         {
             _repo = repo;
         }
@@ -32,18 +32,7 @@ namespace CarouselForBooksApplication.Controllers
         // GET: BookController/Details/5
         public async Task<ActionResult> Details(int id)
         {
-            if (HttpContext.Session.GetString("token") != null)
-            {
-                string token = HttpContext.Session.GetString("token");
-                _repo.GetToken(token);
-                var book = await _repo.GetT(id);
-                if (book != null)
-                {
-                    return View(book);
-                }
-                return NotFound();
-            }
-            return RedirectToAction("Login", "User", new { area = "" });
+            return await GetBookDetails(id);
         }
 
         // GET: BookController/Create
@@ -59,12 +48,18 @@ namespace CarouselForBooksApplication.Controllers
         {
             try
             {
-                book = await _repo.Add(book);
-                if (book != null)
+                if (HttpContext.Session.GetString("token") != null)
                 {
-                    return RedirectToAction(nameof(Index));
+                    string token = HttpContext.Session.GetString("token");
+                    _repo.GetToken(token); 
+                    book = await _repo.Add(book);
+                    if (book != null)
+                    {
+                        return RedirectToAction(nameof(Index));
+                    }
+                    return View();
                 }
-                return View();
+                return RedirectToAction("Login", "User", new { area = "" });
             }
             catch
             {
@@ -74,6 +69,44 @@ namespace CarouselForBooksApplication.Controllers
 
         // GET: BookController/Edit/5
         public async Task<ActionResult> Edit(int id)
+        {
+            return await GetBookDetails(id);
+        }
+
+        // POST: BookController/Edit/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<ActionResult> Edit(Book book)
+        {
+            try
+            {
+                
+                if (HttpContext.Session.GetString("token") != null)
+                {
+                    string token = HttpContext.Session.GetString("token");
+                    _repo.GetToken(token);
+                    book = await _repo.Update(book);
+                    if (book != null)
+                    {
+                        return RedirectToAction(nameof(Index));
+                    }
+                    return View();
+                }
+                return RedirectToAction("Login", "User", new { area = "" });
+            }
+            catch
+            {
+                return View();
+            }
+        }
+
+        // GET: BookController/Delete/5
+        public async Task<ActionResult> Delete(int id)
+        {
+            return await GetBookDetails(id);
+        }
+
+        private async Task<ActionResult> GetBookDetails(int id)
         {
             if (HttpContext.Session.GetString("token") != null)
             {
@@ -89,37 +122,6 @@ namespace CarouselForBooksApplication.Controllers
             return RedirectToAction("Login", "User", new { area = "" });
         }
 
-        // POST: BookController/Edit/5
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit(Book book)
-        {
-            try
-            {
-                book = await _repo.Update(book);
-                if (book != null)
-                {
-                    return RedirectToAction(nameof(Index));
-                }
-                return View();
-            }
-            catch
-            {
-                return View();
-            }
-        }
-
-        // GET: BookController/Delete/5
-        public async Task<ActionResult> Delete(int id)
-        {
-            var book = await _repo.GetT(id);
-            if (book != null)
-            {
-                return View(book);
-            }
-            return NotFound();
-        }
-
         // POST: BookController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -127,12 +129,18 @@ namespace CarouselForBooksApplication.Controllers
         {
             try
             {
-                book = await _repo.Delete(book.Id);
-                if (book != null)
+                if (HttpContext.Session.GetString("token") != null)
                 {
-                    return RedirectToAction(nameof(Index));
+                    string token = HttpContext.Session.GetString("token");
+                    _repo.GetToken(token);
+                    book = await _repo.Delete(book.Id);
+                    if (book != null)
+                    {
+                        return RedirectToAction(nameof(Index));
+                    }
+                    return View();
                 }
-                return View();
+                return RedirectToAction("Login", "User", new { area = "" });
             }
             catch
             {
